price_fetcher/requirements.txt
# price_fetcher/requirements.txt
requests
psycopg2-binary

